refactor: centralize bed state helpers
