+++ .gitignore
@@ -10,4 +10,5 @@ main.js
 !docs/**/*.js
 !postcss.config.js
 !tailwind.config.js
+!src/**/*.js
